refactor: Change table structure to normalize data
Originally players table setup for id, full_name, number of wins, number
of matches. Data about the number of matches played and number of
matches won was stored in the matches table. Meaning the data was not
yet normalized as it should be. A really big hint came when I realized
one of the functions in tournaments.py was updating both tables in order
to report matches.